+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2020-09-10
+
+### Added
+
+-   preflight resets `<button>` padding
+-   preflight resets `<a>` text color and decoration styles
+-   CSSMonster will throw an error when two files have the same name
+-   `autoresolve` config value -- boolean
+
+### Updated
+
+-   npm packages
+
 ## [0.2.1] - 2020-06-05
 
 ### Updated